% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.alpha.R
\name{add.alpha}
\alias{add.alpha}
\title{Add an alpha value to a colour
http://www.magesblog.com/2013/04/how-to-change-alpha-value-of-colours-in.html}
\usage{
add.alpha(col, alpha = 1)
}
\description{
Add an alpha value to a colour
http://www.magesblog.com/2013/04/how-to-change-alpha-value-of-colours-in.html
}
\keyword{internal}

