% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetasGPCM.R
\name{BetasGPCM}
\alias{BetasGPCM}
\title{BetasGPC}
\usage{
BetasGPCM(thetas, nitems, ncatg, constraint, anchor.param = NULL,
  keep.names = FALSE)
}
\arguments{
\item{thetas}{I don't know what these are.}

\item{nitems}{The number of items}

\item{ncatg}{A vector of length nitems containing the number of categories for each item}

\item{anchor.param}{a matrix of the parameters for the anchor set}

\item{constaint}{c("gpcm", "rasch", "1PL")}
}
\description{
This function simply takes a vector of parameters and converts it to a list.
}
\keyword{#@examples}
\keyword{#@export}
\keyword{BetasGPCM()}
\keyword{add}
\keyword{here}
\keyword{one}
