% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular.indicadores.2.timings.R
\name{calcular.indicadores.2.timings}
\alias{calcular.indicadores.2.timings}
\title{Calculates specificity and sensitivity}
\usage{
calcular.indicadores.2.timings(
  i.current,
  i.timing.1,
  i.timing.2,
  i.timing.absolute = FALSE,
  i.timing.labels,
  i.output = ".",
  i.graph.title = "",
  i.graph.file = FALSE,
  i.graph.file.name = "",
  i.mem.info = TRUE,
  i.labels.axis = c("Week", "Weekly rate"),
  i.labels.periods = c("Pre", "Epidemic", "Post"),
  i.labels.details = c("Sensitivity", "Specificity",
    "First row represents real and second row observed pos/neg values")
)
}
\description{
Calculates specificity and sensitivity
}
\keyword{internal}
