% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular.indicadores.R
\name{calcular.indicadores}
\alias{calcular.indicadores}
\title{Calculates specificity and sensitivity}
\usage{
calcular.indicadores(
  i.current,
  i.umbral.pre,
  i.umbral.pos = NA,
  i.intensidades,
  i.duracion.intensidad = 10,
  i.equal = FALSE,
  i.metodo.calculo = "default",
  i.semanas.por.encima = 1,
  i.valores.parametro.deteccion = seq(0.1, 5, 0.1),
  i.output = ".",
  i.graph = FALSE,
  i.graph.name = "",
  i.mem.info = TRUE,
  i.labels.axis = c("Week", "Weekly rate"),
  i.labels.periods = c("Pre", "Epidemic", "Post"),
  i.labels.intensities = c("Epidemic thr", "Medium thr", "High thr", "Very high thr"),
  i.labels.details = c("algorithm", "threshold", "Method used",
    "weeks above/below the threshold", "week(s) above the threshold", "Sensitivity",
    "Specificity")
)
}
\description{
Calculates specificity and sensitivity
}
\keyword{internal}
