% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular.optimo.criterio.R
\name{calcular.optimo.criterio}
\alias{calcular.optimo.criterio}
\title{calculates optimum: fixed criteria for slope}
\usage{
calcular.optimo.criterio(i.curva.map, i.criterio = 2.8)
}
\description{
calculates optimum: fixed criteria for slope
}
\keyword{internal}
