% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular.optimo.derivada.R
\name{calcular.optimo.derivada}
\alias{calcular.optimo.derivada}
\title{calculates optimum: second derivative equals 0 (change signs from - to +, or + to -)}
\usage{
calcular.optimo.derivada(i.curva.map)
}
\description{
calculates optimum: second derivative equals 0 (change signs from - to +, or + to -)
}
\keyword{internal}
