% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular.optimo.original.R
\name{calcular.optimo.original}
\alias{calcular.optimo.original}
\title{calculates optimum: original method: second derivative + axis change}
\usage{
calcular.optimo.original(i.curva.map, i.hsuav)
}
\description{
calculates optimum: original method: second derivative + axis change
}
\keyword{internal}
