% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular.optimo.pendiente.R
\name{calcular.optimo.pendiente}
\alias{calcular.optimo.pendiente}
\title{calculates optimum: slope matches the overall slope}
\usage{
calcular.optimo.pendiente(i.curva.map)
}
\description{
calculates optimum: slope matches the overall slope
}
\keyword{internal}
