% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular.optimo.R
\name{calcular.optimo}
\alias{calcular.optimo}
\title{calculates optimum generic function}
\usage{
calcular.optimo(i.curva.map, i.metodo, i.parametro)
}
\description{
calculates optimum generic function
}
\keyword{internal}
