% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular.tickmarks.R
\name{calcular.tickmarks}
\alias{calcular.tickmarks}
\title{Calculate optimal tickmarks for a graph}
\usage{
calcular.tickmarks(maximo.y, ticks.objetivo = 10, ticks.validos = c(1, 2, 5,
  10, 20, 25, 50, 100, 250, 500, 1000, 1500, 2000, 2500, 5000))
}
\description{
Calculate optimal tickmarks for a graph
}
\keyword{internal}

