% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparar.metodos.R
\name{comparar.metodos}
\alias{comparar.metodos}
\title{Compares outputs from two methods of locating the epidemic}
\usage{
comparar.metodos(i.data)
}
\description{
Compares outputs from two methods of locating the epidemic
}
\keyword{internal}
