% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_meas.R
\name{conv_meas}
\alias{conv_meas}
\title{conv_meas Calculate the convex hull for both up and down being the "inside" of the curve and take the ratio 
of the path lengths. A value < 1 indicates convex (concave up), while a value > 1 indicates concave (concave down). 
A value exactly 1 indicates a line (or that it is no more convex than concave).}
\usage{
conv_meas(x, y)
}
\description{
conv_meas Calculate the convex hull for both up and down being the "inside" of the curve and take the ratio 
of the path lengths. A value < 1 indicates convex (concave up), while a value > 1 indicates concave (concave down). 
A value exactly 1 indicates a line (or that it is no more convex than concave).
}
\keyword{internal}
