% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.calendar.R
\name{create.calendar}
\alias{create.calendar}
\title{create epidemiological calendar}
\usage{
create.calendar(
  i.ini = as.POSIXct("1980-01-01"),
  i.fin = as.POSIXct(paste(format(Sys.Date(), "\%Y"), "12", "31", sep = "-"))
)
}
\description{
create epidemiological calendar
}
\keyword{internal}
