% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrfGPCM.R
\name{CrfGPCM}
\alias{CrfGPCM}
\title{Cumulative Response Function GPCM}
\usage{
CrfGPCM(betas, z, IRT.param = TRUE, log = FALSE,
  eps = .Machine$double.eps^(1/2))
}
\arguments{
\item{betas}{This is a list of betas from the BetasGPCM function}

\item{z}{I am not sure what z does.}

\item{IRT.param}{Should IRT parameterization be used? TRUE/FALSE}
}
\description{
This calculates the CRF of the GPCM.
}
\keyword{#@examples}
\keyword{#@export}
\keyword{CrfGPCM()}
\keyword{add}
\keyword{here}
\keyword{one}
