% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimem.R
\name{epimem}
\alias{epimem}
\title{Deprecated function(s) in the mem package}
\usage{
epimem(
  i.data,
  i.tails = 1,
  i.levels = c(0.4, 0.9, 0.975),
  i.type = 2,
  i.level = 0.95,
  ...
)
}
\arguments{
\item{i.data}{Original parameter deprecated in the new version}

\item{i.tails}{Original parameter deprecated in the new version}

\item{i.levels}{Original parameter deprecated in the new version}

\item{i.type}{Original parameter deprecated in the new version}

\item{i.level}{Original parameter deprecated in the new version}

\item{...}{Parameters (that have not changed) to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the mem package.  They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
  \code{epimem} \tab now a synonym for \code{\link{memmodel}}
}
}

