% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitiming.R
\name{epitiming}
\alias{epitiming}
\title{Deprecated function(s) in the mem package}
\usage{
epitiming(i.data, ...)
}
\arguments{
\item{i.data}{Original parameter deprecated in the new version}

\item{...}{Parameters (that have not changed) to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the mem package.  They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
  \code{epitiming} \tab now a synonym for \code{\link{memtiming}}
}
}

