% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer.datos.post.epi.R
\name{extraer.datos.post.epi}
\alias{extraer.datos.post.epi}
\title{Extract post-epidemic rates}
\usage{
extraer.datos.post.epi(i.epi)
}
\description{
Extract post-epidemic rates
}
\keyword{internal}

