% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.missing.R
\name{fill.missing}
\alias{fill.missing}
\title{fills in missing values inside the season with smoothing regression}
\usage{
fill.missing(i.datos)
}
\description{
fills in missing values inside the season with smoothing regression
}
\keyword{internal}
