% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.epiweek.R
\name{get.epiweek}
\alias{get.epiweek}
\title{create epidemiological calendar}
\usage{
get.epiweek(i.date = as.POSIXct(as.POSIXlt(Sys.Date()), tz = "CET"))
}
\description{
create epidemiological calendar
}
\keyword{internal}
