% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHpoints.R
\name{GHpoints}
\alias{GHpoints}
\title{Gaussian-Hermite Points}
\usage{
GHpoints(form, k)
}
\arguments{
\item{form}{form}

\item{k}{The number of points}
}
\description{
Gets the Gaussian-Hermite points for a paticular form and size.
}
\keyword{#@examples}
\keyword{#@export}
\keyword{GHpoints()}
\keyword{add}
\keyword{here}
\keyword{one}
