% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GpcmMEM.R
\name{GpcmMEM}
\alias{GpcmMEM}
\title{MEM Estimation for the GPCM}
\usage{
GpcmMEM(data, anchor, anchor.param = NULL, constraint = c("gpcm", "1PL",
  "rasch"), hess = F, IRT.param = TRUE, start.val = NULL,
  na.action = NULL, trace = T, control = list())
}
\arguments{
\item{data}{A matrix of response}

\item{anchor}{A vector indicating which items are anchor items}

\item{anchor.param}{The previous item estimates for the ancor items}

\item{constraint}{c("gpcm","1PL","rasch")}

\item{hess}{A logical indicating whether or not to calculate and return the hessian (for standard errors)}

\item{IRT.param}{Logical indicating whether to use IRT parameterization}

\item{trace}{Interger specifcying progress of optimization routine}
}
\description{
This function will allow you to calibrate and link IRT models for the GPCM.
}
\keyword{add}
\keyword{here}
\keyword{one}
