% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconfianza.aritmetica.R
\name{iconfianza.aritmetica}
\alias{iconfianza.aritmetica}
\title{confidence interval for the arithmetic mean using the normal approximation}
\usage{
iconfianza.aritmetica(datos, nivel = 0.95, ic = T, colas = 2)
}
\description{
confidence interval for the arithmetic mean using the normal approximation
}
\keyword{internal}
