% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconfianza.completo.R
\name{iconfianza.completo}
\alias{iconfianza.completo}
\title{Int. Confianza de 2 y 1 cola, por encima y por debajo. De la media y el punto.}
\usage{
iconfianza.completo(i.datos, tipo = 1, i.nivel = 0.95)
}
\description{
Int. Confianza de 2 y 1 cola, por encima y por debajo. De la media y el punto.
}
\keyword{internal}
