% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconfianza.logx.R
\name{iconfianza.logx}
\alias{iconfianza.logx}
\title{confidence interval for a point (using log transformation and geometric mean)}
\usage{
iconfianza.logx(datos, nivel = 0.95, ic = TRUE, colas = 2, use.t = FALSE)
}
\description{
confidence interval for a point (using log transformation and geometric mean)
}
\keyword{internal}
