% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconfianza.percentil.eqnpar.R
\name{iconfianza.percentil.eqnpar}
\alias{iconfianza.percentil.eqnpar}
\title{confidence interval por the median using KC Method}
\usage{
iconfianza.percentil.eqnpar(datos, q = 0.5, nivel = 0.95, ic = T, colas = 2)
}
\description{
confidence interval por the median using KC Method
}
\keyword{internal}
