% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconfianza.R
\name{iconfianza}
\alias{iconfianza}
\title{generic confidence interval calculation function}
\usage{
iconfianza(datos, nivel = 0.95, tipo = 1, ic = T, tipo.boot = "normal",
  iteraciones.boot = 10000, colas = 2)
}
\description{
generic confidence interval calculation function
}
\keyword{internal}
