% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinpredGPCM.R
\name{LinpredGPCM}
\alias{LinpredGPCM}
\title{Linear predictor for the GPCM}
\usage{
LinpredGPCM(betas, z, IRT.param = TRUE)
}
\arguments{
\item{betas}{Item parameters for a particular}

\item{z}{a particular theta value
#@export
#@examples
LinpredGPCM()}
}
\description{
Takes betas and z to calculate the linear combination of a response category for the GPCM
}
