% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLikGPCM.R
\name{LogLikGPCM}
\alias{LogLikGPCM}
\title{GPCM Log-likelihood function}
\usage{
LogLikGPCM(thetas, constraint, anchor = NULL)
}
\arguments{
\item{thetas}{}

\item{constraint}{c("gpcm","rasch","1PL")}

\item{anchor}{fixed values of the anchor set}
}
\description{
Calculates the log-likelihood given a set of thetas.
}
\keyword{#@examples}
\keyword{#@export}
\keyword{LogLikGPCM()}
\keyword{add}
\keyword{here}
\keyword{one}
