% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.n.valores.R
\name{max.n.valores}
\alias{max.n.valores}
\title{highest n values of a data set, removing Inf and -Inf}
\usage{
\method{max}{n.valores}(mis.datos, n.max = 1)
}
\description{
highest n values of a data set, removing Inf and -Inf
}
\keyword{internal}
