% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnvalores.R
\name{maxnvalores}
\alias{maxnvalores}
\title{highest n values of a data set, removing Inf and -Inf}
\usage{
maxnvalores(mis.datos, n.max = 1)
}
\description{
highest n values of a data set, removing Inf and -Inf
}
\keyword{internal}
