\name{mem-internal}
\alias{calcular.map}
\alias{calcular.optimo.criterio}
\alias{calcular.optimo.derivada}
\alias{calcular.optimo.original}
\alias{calcular.optimo.pendiente}
\alias{calcular.optimo.varadhan}
\alias{calcular.optimo}
\alias{max.fix.na}
\alias{max.n.valores}
\alias{min.fix.na}
\alias{min.n.valores}
\alias{normalizar}
\alias{output.ci}
\alias{suavizado}
\alias{missings.inside}
\alias{fill.missing}
\alias{iconfianza}
\alias{iconfianza.aritmetica}
\alias{iconfianza.geometrica}
\alias{iconfianza.x}
\alias{iconfianza.logx}
\alias{iconfianza.percentil.kc}
\alias{iconfianza.percentil.boot}
\alias{extraer.datos.optimo.map}
\alias{extraer.datos.pre.epi}
\alias{extraer.datos.post.epi}
\alias{extraer.datos.curva.map}
\alias{semana.absoluta}
\alias{localizar.epidemia}
\alias{influenza}
\docType{package}
\title{
Internal functions for mem package
}
\description{
These are internal functions used by the mem package and for use of the program, it cant be used by the user since they have not been designed for general use.
}

