% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\docType{package}
\name{mem-package}
\alias{mem}
\alias{mem-package}
\title{mem: The Moving Epidemic Method}
\description{
The Moving Epidemic Method, created by T Vega and JE Lozano (2012, 2015) \doi{10.1111/j.1750-2659.2012.00422.x}, \doi{10.1111/irv.12330}, allows the weekly assessment of the epidemic and intensity status to help in routine respiratory infections surveillance in health systems. Allows the comparison of different epidemic indicators, timing and shape with past epidemics and across different regions or countries with different surveillance systems. Also, it gives a measure of the performance of the method in terms of sensitivity and specificity of the alert week.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lozalojo/mem}
  \item Report bugs at \url{https://github.com/lozalojo/mem/issues}
}

}
\author{
\strong{Maintainer}: Jose E. Lozano \email{lozalojo@gmail.com}

Authors:
\itemize{
  \item Jose E. Lozano \email{lozalojo@gmail.com}
}

}
\keyword{internal}
