% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memsurveillance.animated.R
\name{memsurveillance.animated}
\alias{memsurveillance.animated}
\title{Creates the animated surveillance graph of the current season}
\usage{
memsurveillance.animated(i.current, i.epidemic.thresholds = NA,
  i.intensity.thresholds = NA, i.output = ".", i.graph.file.name = "",
  i.delay = 100, i.loop = 0, ...)
}
\arguments{
\item{i.current}{Current season weekly rates.}

\item{i.epidemic.thresholds}{Pre and post epidemic threholds.}

\item{i.intensity.thresholds}{Intensity thresholds.}

\item{i.output}{Directory where graph is saved.}

\item{i.graph.file.name}{Name of the graph.}

\item{i.delay}{Delay between frames of the animated gif.}

\item{i.loop}{Number of loops for the animated dif, 0 for Infinite.}

\item{...}{Additional parameters parsed to memsurveillance.}
}
\value{
\code{memsurveillance.animated} writes a gif graph of the surveillance of this season.
}
\description{
Function \code{memsurveillance.animated} creates an animated surveillance graph for the current season.
}
\details{
Input data must be the current season and an object of class \code{mem}. The output
graph contains the weekly rates series along with the epidemic and intensity threshols
located at the exact situation where the epidemic started. If there is no epidemic yet,
only the epidemic threshold is placed.
}
\examples{
# Castilla y Leon Influenza Rates data
data(flucyl)
# Data of the last season
cur<-flucyl[8]
# The model
epi<-memmodel(flucyl[1:7])
# Epidemic thresholds
e.thr<-epi$epidemic.thresholds
# Intensity threhsolds
i.thr<-epi$intensity.thresholds
# Set the working directory to whererever you want to store the graph file
setwd(".")
m1<-a<-memsurveillance.animated(cur, i.graph.file.name="Animated",i.epidemic.thresholds = e.thr,
i.intensity.thresholds = i.thr, i.pos.epidemic = TRUE)

}
\author{
Jose E. Lozano \email{lozalojo@gmail.com}
}
\references{
Vega T., Lozano J.E. (2004) Modelling influenza epidemic - can we detect the beginning
and predict the intensity and duration? International Congress Series 1263 (2004)
281-283.\cr
Vega T., Lozano J.E. (2012) Influenza surveillance in Europe: establishing epidemic
thresholds by the Moving Epidemic Method. Influenza and Other Respiratory Viruses,
DOI:10.1111/j.1750-2659.2012.00422.x.
}
\keyword{influenza}

