% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memsurveillance.R
\name{memsurveillance}
\alias{memsurveillance}
\title{Creates the surveillance graph of the current season}
\usage{
memsurveillance(i.current, i.flu, i.output = ".", i.graph.name = "",
  i.week.report = NA, i.equal = F, i.pos.epidemic = F,
  i.no.epidemic = F, i.no.intensity = F, i.epidemic.start = NA,
  i.range.x = c(40, 20), i.range.x.53 = F, i.range.y = NA,
  i.no.labels = F, i.start.end.marks = T)
}
\arguments{
\item{i.current}{Current season weekly rates.}

\item{i.flu}{Object of class \code{mem} used to calculate thresholds.}

\item{i.output}{Directory where graph is saved.}

\item{i.graph.name}{Name of the graph.}

\item{i.week.report}{Week to use in the report.}

\item{i.equal}{If post epidemic and preepidemic thresholds must be equal (force post epidemic to be equal to the pre epidemic threshold).}

\item{i.pos.epidemic}{Print post epidemic threhsold.}

\item{i.no.epidemic}{Force no start of the epidemic, print only the epidemic threshold.}

\item{i.no.intensity}{Do not print intensity threholds.}

\item{i.epidemic.start}{Week to force start of the epidemic.}

\item{i.range.x}{Range of weeks.}

\item{i.range.x.53}{Is there a week 53 this season.}

\item{i.range.y}{Range of graph.}

\item{i.no.labels}{Do not use labels.}

\item{i.start.end.marks}{Do not place start and end marks of the epidemic.}
}
\value{
\code{memsurveillance} writes a tiff graph of the surveillance of this season.
}
\description{
Function \code{memsurveillance} creates a surveillance graph for the current season.
}
\details{
Input data must be the current season and an object of class \code{mem}. The output
graph contains the weekly rates series along with the epidemic and intensity threshols
located at the exact situation where the epidemic started. If there is no epidemic yet,
only the epidemic threshold is placed.
}
\examples{
# Castilla y Leon Influenza Rates data
data(flucyl)
# Data of the last season
cur<-flucyl[8]
# The model
epi<-memmodel(flucyl[1:7])
# Set the working directory to whererever you want to store the graph file
setwd(".")
# The graph, default values
memsurveillance(cur,epi,i.graph.name="graph 1")
# No intensity levels
memsurveillance(cur,epi,i.graph.name="graph 2",i.no.intensity=TRUE)
# No start/end tickmarks
memsurveillance(cur,epi,i.graph.name="graph 3",i.start.end.marks=FALSE)
# Post-epidemic threshold
memsurveillance(cur,epi,i.graph.name="graph 4",i.pos.epidemic=TRUE)
# Report for week 2, instead of all data
memsurveillance(cur,epi,i.graph.name="graph 5",i.week.report=2)

}
\author{
Jose E. Lozano \email{lozalojo@gmail.com}
}
\references{
Vega T., Lozano J.E. (2004) Modelling influenza epidemic - can we detect the beginning
and predict the intensity and duration? International Congress Series 1263 (2004)
281-283.\cr
Vega T., Lozano J.E. (2012) Influenza surveillance in Europe: establishing epidemic
thresholds by the Moving Epidemic Method. Influenza and Other Respiratory Viruses,
DOI:10.1111/j.1750-2659.2012.00422.x.
}
\keyword{influenza}

