% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min.n.valores.R
\name{min.n.valores}
\alias{min.n.valores}
\title{lowest n values of a data set, removing Inf and -Inf}
\usage{
\method{min}{n.valores}(mis.datos, n.min = 1)
}
\description{
lowest n values of a data set, removing Inf and -Inf
}
\keyword{internal}
