% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minnvalores.R
\name{minnvalores}
\alias{minnvalores}
\title{lowest n values of a data set, removing Inf and -Inf}
\usage{
minnvalores(mis.datos, n.min = 1)
}
\description{
lowest n values of a data set, removing Inf and -Inf
}
\keyword{internal}
