% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimum.by.inspection.R
\name{optimum.by.inspection}
\alias{optimum.by.inspection}
\title{Inspection calcultation of the optimum}
\usage{
optimum.by.inspection(i.data, i.detection.values = seq(1.5, 4.5, 0.1))
}
\arguments{
\item{i.data}{Data frame of input data.}

\item{i.detection.values}{values to use in the i.param value of \code{memtiming}.}
}
\value{
\code{optimum.by.inspection} returns a list.
An object of class \code{mem} is a list containing at least the following components:
  \item{optimum}{optimum value.}
  \item{optimum.data}{Data related to the optimum value.}
  \item{summary.data}{Data for all values tested.}
  \item{inspection.data}{Detailed results of each iteration.}
}
\description{
Function \code{optimum.by.inspection} perform an analysis of mem parameters
to find the one that fits better a panel of experts inspection criterium.
}
\details{
To be written
}
\examples{
# Castilla y Leon Influenza Rates data
data(flucyl)
# Inspection. It runs interactively (uncomment to run)
#opt.ins<-optimum.by.inspection(flucyl,i.detection.values=seq(2.0,3.0,0.1))
#opt.ins$optimum.data

}
\author{
Jose E. Lozano \email{lozalojo@gmail.com}
}
\references{
Vega T., Lozano J.E. (2004) Modelling influenza epidemic - can we detect the beginning
and predict the intensity and duration? International Congress Series 1263 (2004)
281-283.\cr
Vega T., Lozano J.E. (2012) Influenza surveillance in Europe: establishing epidemic
thresholds by the Moving Epidemic Method. Influenza and Other Respiratory Viruses,
DOI:10.1111/j.1750-2659.2012.00422.x.
}
\keyword{influenza}

