% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer.verdadero.R
\name{primer.verdadero}
\alias{primer.verdadero}
\title{find first TRUE value function, returning NA if not present}
\usage{
primer.verdadero(x)
}
\description{
find first TRUE value function, returning NA if not present
}
\keyword{internal}
