% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPlots.R
\name{processPlots}
\alias{processPlots}
\title{Full process plots for mem}
\usage{
processPlots(i.flu, i.output = ".", i.prefix = "")
}
\arguments{
\item{i.flu}{Object of class flu.}

\item{i.output}{Output directory.}

\item{i.prefix}{Prefix for all files to be output.}
}
\value{
\code{processPlots} prints a set of graphs.
}
\description{
Function \code{processPlots} creates graphs of all mem process
}
\details{
To be written
}
\examples{
# Castilla y Leon Influenza Rates data
data(flucyl)
# Graphs
epi<-memmodel(flucyl)
processPlots(epi)

}
\author{
Jose E. Lozano \email{lozalojo@gmail.com}
}
\references{
Vega T., Lozano J.E. (2004) Modelling influenza epidemic - can we detect the beginning
and predict the intensity and duration? International Congress Series 1263 (2004)
281-283.\cr
Vega T., Lozano J.E. (2012) Influenza surveillance in Europe: establishing epidemic
thresholds by the Moving Epidemic Method. Influenza and Other Respiratory Viruses,
DOI:10.1111/j.1750-2659.2012.00422.x.
}
\keyword{influenza}

