% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_convrate.R
\name{roll_convrate}
\alias{roll_convrate}
\title{roll_convrate is used by percentage.added}
\usage{
roll_convrate(i.data, i.n)
}
\description{
roll_convrate is used by percentage.added
}
\keyword{internal}
