% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semana.absoluta.R
\name{semana.absoluta}
\alias{semana.absoluta}
\title{Transforms relative weeks to absolute weeks in a 1-52 normal season}
\usage{
semana.absoluta(i.semana, i.semana.inicio)
}
\description{
Transforms relative weeks to absolute weeks in a 1-52 normal season
}
\keyword{internal}
