% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartValGPCM .R
\name{StartValGPCM}
\alias{StartValGPCM}
\title{Starting Value Generation for GPCM}
\usage{
StartValGPCM(start.val, data, weight, constraint, ncatg, IRT.param)
}
\arguments{
\item{start.val}{an arbitrary vector fitting the format for the dataset}

\item{data}{The dataset for which to generate starting values}

\item{weight}{weighting}

\item{ncatg}{a vector indicating how many categories each item has}

\item{IRT.param}{TRUE or FALSE}
}
\description{
This function generates starting values for the GPCM.
}
\keyword{#@examples}
\keyword{#@export}
\keyword{StartValGPCM()}
\keyword{add}
\keyword{here}
\keyword{one}
