% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformdata.back.R
\name{transformdata.back}
\alias{transformdata.back}
\title{Tranform the historial series from the mem default input format (weeksxseasons)
to a single serie data.frame, with season, week and rate column}
\usage{
transformdata.back(i.data, i.name = "rates", i.range.x = c(1, 53))
}
\description{
Tranform the historial series from the mem default input format (weeksxseasons)
to a single serie data.frame, with season, week and rate column
}
\keyword{internal}

