% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformdata.R
\name{transformdata}
\alias{transformdata}
\title{Data transformation}
\usage{
transformdata(i.data, i.week.first = 40, i.week.last = 20)
}
\arguments{
\item{i.data}{Data frame of input data.}

\item{i.week.first}{First surveillance week.}

\item{i.week.last}{Last surveillance week.}
}
\value{
\code{transformdata} returns a data.frame where each column has a different season and
rownames are the name of the epidemiological week.
}
\description{
Function \code{transformdata} transforms data from year,week,rate to week,rate1,...,rateN
suitable to use with mem.
}
\details{
Yet to be written
}
\examples{
# Castilla y Leon Influenza Rates data
data(flucylraw)
# Transform data
newdata<-transformdata(flucylraw)$tdata
epi<-memmodel(newdata)
print(epi)
summary(epi)
plot(epi)

}
\author{
Jose E. Lozano \email{lozalojo@gmail.com}
}
\references{
Vega T., Lozano J.E. (2004) Modelling influenza epidemic - can we detect the beginning
and predict the intensity and duration? International Congress Series 1263 (2004)
281-283.\cr
Vega T., Lozano J.E. (2012) Influenza surveillance in Europe: establishing epidemic
thresholds by the Moving Epidemic Method. Influenza and Other Respiratory Viruses,
DOI:10.1111/j.1750-2659.2012.00422.x.
}
\keyword{influenza}

