% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformseries.loess.R
\name{transformseries.loess}
\alias{transformseries.loess}
\title{loess transformation}
\usage{
transformseries.loess(i.x, i.positive = FALSE, ...)
}
\description{
loess transformation
}
\keyword{internal}
