% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformseries.moving.average.R
\name{transformseries.moving.average}
\alias{transformseries.moving.average}
\title{moving average transformation}
\usage{
transformseries.moving.average(x, i.number = 3)
}
\description{
moving average transformation
}
\keyword{internal}
