% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformseries.odd.R
\name{transformseries.odd}
\alias{transformseries.odd}
\title{fills in missing values inside the season with smoothing regression}
\usage{
transformseries.odd(x, mult = 0)
}
\description{
fills in missing values inside the season with smoothing regression
}
\keyword{internal}
