% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformseries.R
\name{transformseries}
\alias{transformseries}
\title{Transformation of series of data}
\usage{
transformseries(i.data, i.transformation = 1)
}
\arguments{
\item{i.data}{Historical data series.}

\item{i.transformation}{Transformation to apply to the dataset.}
}
\value{
\code{transformseries} The transformed dataset.
}
\description{
Function \code{transformseries} transforms whole datasets.
}
\details{
Input data must be a data.frame with each column a surveillance season and each
row a week.

Transformation options:

\tabular{rlll}{
\tab [1] \tab No transformation\cr
\tab [2] \tab Odd\cr
\tab [3] \tab Fill missing data\cr
\tab [4] \tab Loess\cr
\tab [5] \tab Two waves (observed)\cr
\tab [6] \tab Two waves (expected)\cr
}
}
\examples{
# Castilla y Leon Influenza Rates data
data(flucyl)
# Data of the last season
transformseries(flucyl,2)

}
\references{
Vega Alonso, Tomas, Jose E Lozano Alonso, Raul Ortiz de Lejarazu, and Marisol Gutierrez Perez. 2004.
Modelling Influenza Epidemic: Can We Detect the Beginning and Predict the Intensity and Duration?
International Congress Series, Options for the Control of Influenza V. Proceedings of the International
Conference on Options for the Control of Influenza V, 1263 (June): 281-83. doi:10.1016/j.ics.2004.02.121.\cr
Vega, Tomas, Jose Eugenio Lozano, Tamara Meerhoff, Rene Snacken, Joshua Mott, Raul Ortiz de Lejarazu, and
Baltazar Nunes. 2013. Influenza Surveillance in Europe: Establishing Epidemic Thresholds by the Moving
Epidemic Method. Influenza and Other Respiratory Viruses 7 (4): 546-58. doi:10.1111/j.1750-2659.2012.00422.x.\cr
Vega, Tomas, Jose E. Lozano, Tamara Meerhoff, Rene Snacken, Julien Beaute, Pernille Jorgensen, Raul Ortiz
de Lejarazu, et al. 2015. Influenza Surveillance in Europe: Comparing Intensity Levels Calculated Using
the Moving Epidemic Method. Influenza and Other Respiratory Viruses 9 (5): 234-46. doi:10.1111/irv.12330.
}
\author{
Jose E. Lozano \email{lozalojo@gmail.com}
}
\keyword{influenza}
