% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformseries.spline.R
\name{transformseries.spline}
\alias{transformseries.spline}
\title{spline transformation}
\usage{
transformseries.spline(i.x, i.positive = F, ...)
}
\description{
spline transformation
}
\keyword{internal}
