% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full.series.graph.R
\name{full.series.graph}
\alias{full.series.graph}
\title{Creates the historical series graph of the datasets}
\usage{
full.series.graph(i.data, i.range.x = c(30, 29), i.range.y = NA,
  i.output = ".", i.graph.title = "", i.graph.subtitle = "",
  i.graph.file = T, i.graph.file.name = "", i.plot.timing = F,
  i.plot.intensity = F, i.color.pattern = c("#C0C0C0", "#606060", "#000000",
  "#808080", "#000000", "#001933", "#00C000", "#800080", "#FFB401", "#8c6bb1",
  "#88419d", "#810f7c", "#4d004b"), ...)
}
\arguments{
\item{i.data}{Historical data series.}

\item{i.range.x}{Range x (surveillance weeks) of graph.}

\item{i.range.y}{Range y of graph.}

\item{i.output}{Directory where graph is saved.}

\item{i.graph.title}{Title of the graph.}

\item{i.graph.subtitle}{Subtitle of the graph.}

\item{i.graph.file}{Graph to a file.}

\item{i.graph.file.name}{Name of the graph.}

\item{i.plot.timing}{Plot the timing of epidemics.}

\item{i.plot.intensity}{Plot the intensity levels.}

\item{i.color.pattern}{colors to use in the graph.}

\item{...}{other parameters passed to memmodel.}
}
\value{
\code{full.series.graph} writes a tiff graph of the full series of the dataset.

Color codes:
1: Axis.
2: Tickmarks.
3: Axis labels.
4: Series line.
5: Series dots (default).
6: Title and subtitle.
7: Series dots (pre-epidemic).
8: Series dots (epidemic).
9: Series dots (post-epidemic).
10: Epidemic threshold.
11: Medium threshold.
12: High threshold.
13: Very high threshold.
}
\description{
Function \code{full.series.graph} creates a graph with the whole data.
}
\details{
Input data must be a data.frame with each column a surveillance season and each
row a week.
}
\examples{
# Castilla y Leon Influenza Rates data
data(flucyl)
# Data of the last season
full.series.graph(flucyl)

}
\author{
Jose E. Lozano \email{lozalojo@gmail.com}
}
\references{
Vega T., Lozano J.E. (2004) Modelling influenza epidemic - can we detect the beginning
and predict the intensity and duration? International Congress Series 1263 (2004)
281-283.\cr
Vega T., Lozano J.E. (2012) Influenza surveillance in Europe: establishing epidemic
thresholds by the Moving Epidemic Method. Influenza and Other Respiratory Viruses,
DOI:10.1111/j.1750-2659.2012.00422.x.
}
\keyword{influenza}

