% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconfianza.geometrica.R
\name{iconfianza.geometrica}
\alias{iconfianza.geometrica}
\title{confidence interval for the geometric mean using the log-normal approximation}
\usage{
iconfianza.geometrica(datos, nivel = 0.95, ic = T, colas = 2)
}
\description{
confidence interval for the geometric mean using the log-normal approximation
}
\keyword{internal}
