% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconfianza.x.R
\name{iconfianza.x}
\alias{iconfianza.x}
\title{confidence interval for a point (using arithmetic mean)}
\usage{
iconfianza.x(datos, nivel = 0.95, ic = T, colas = 2)
}
\description{
confidence interval for a point (using arithmetic mean)
}
\keyword{internal}
