% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage.added.R
\name{percentage.added}
\alias{percentage.added}
\title{For use with transformseries.multiple}
\usage{
percentage.added(i.data, n)
}
\description{
For use with transformseries.multiple
}
\keyword{internal}
