% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.analysis.R
\name{roc.analysis}
\alias{roc.analysis}
\title{ROC analysis to find optimum parameter value}
\usage{
roc.analysis(i.data, i.seasons = 10, i.type.threshold = 5,
  i.level.threshold = 0.95, i.tails.threshold = 1, i.type.intensity = 6,
  i.level.intensity = c(0.4, 0.9, 0.975), i.tails.intensity = 1,
  i.type.curve = 2, i.level.curve = 0.95, i.type.other = 2,
  i.level.other = 0.95, i.method = 2, i.param.values = seq(2, 3, 0.1),
  i.n.max = -1, i.type.boot = "norm", i.iter.boot = 10000,
  i.calculation.method = "default", i.goodness.method = "cross",
  i.detection.values = seq(2, 3, 0.1), i.weeks.above = 1, i.output = ".",
  i.graph = F, i.prefix = "")
}
\arguments{
\item{i.param.values}{range of i.param values to test.}

\item{...}{same as \code{memgoodness}}
}
\value{
\code{roc.analysis} returns a list.
An object of class \code{mem} is a list containing at least the following components:
  \item{optimum}{optimum value.}
  \item{results}{Detailed results of each iteration.}
}
\description{
Function \code{roc.analysis} perform a ROC analysis
}
\details{
To be written
}
\examples{
# Castilla y Leon Influenza Rates data
data(flucyl)
# Finds the timing of the first season: 2001/2002
repi<-roc.analysis(flucyl,i.param.values=seq(2.5,2.8,0.1),i.detection.values=seq(2.5,2.8,0.1))
repi

}
\author{
Jose E. Lozano \email{lozalojo@gmail.com}
}
\references{
Vega T., Lozano J.E. (2004) Modelling influenza epidemic - can we detect the beginning
and predict the intensity and duration? International Congress Series 1263 (2004)
281-283.\cr
Vega T., Lozano J.E. (2012) Influenza surveillance in Europe: establishing epidemic
thresholds by the Moving Epidemic Method. Influenza and Other Respiratory Viruses,
DOI:10.1111/j.1750-2659.2012.00422.x.
}
\keyword{influenza}

