% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.analysis.R
\name{roc.analysis}
\alias{roc.analysis}
\title{ROC analysis to find optimum parameter value}
\usage{
roc.analysis(i.data, i.param.values = seq(1.5, 4.5, 0.1), i.prefix.roc = "",
  ...)
}
\arguments{
\item{i.data}{Data frame of input data.}

\item{i.param.values}{range of i.param values to test.}

\item{i.prefix.roc}{prefix used for naming graphs.}

\item{...}{other paramaters to be used by memgoodness function.}
}
\value{
\code{roc.analysis} returns a list.
An object of class \code{mem} is a list containing at least the following components:
  \item{optimum}{optimum value.}
  \item{results}{Detailed results of each iteration.}
}
\description{
Function \code{roc.analysis} perform a ROC analysis
}
\details{
To be written
}
\examples{
# Castilla y Leon Influenza Rates data
data(flucyl)
# ROC analysis
epi.roc<-roc.analysis(flucyl,i.param.values=seq(2.5,2.8,0.1),i.detection.values=seq(2.5,2.8,0.1))
epi.roc$results

}
\author{
Jose E. Lozano \email{lozalojo@gmail.com}
}
\references{
Vega T., Lozano J.E. (2004) Modelling influenza epidemic - can we detect the beginning
and predict the intensity and duration? International Congress Series 1263 (2004)
281-283.\cr
Vega T., Lozano J.E. (2012) Influenza surveillance in Europe: establishing epidemic
thresholds by the Moving Epidemic Method. Influenza and Other Respiratory Viruses,
DOI:10.1111/j.1750-2659.2012.00422.x.
}
\keyword{influenza}

