% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformseries.multiple.R
\name{transformseries.multiple}
\alias{transformseries.multiple}
\title{divide a series in parts}
\usage{
transformseries.multiple(
  i.data,
  i.max.epidemic.duration = NA,
  i.max.season.duration = NA,
  i.waves = NA,
  i.waves.range = NA,
  i.intra.param = 3,
  i.inter.param = 2,
  i.min.separation = 1,
  i.output = NA,
  i.prefix = "Multiple waves",
  i.force.smooth = FALSE,
  i.split.top = 3,
  i.param = 2.8,
  i.p1titles = c("Series and smooth", "Data to be used", "Week", "Data"),
  i.p2titles = c("Iteration", "Week", "Data"),
  i.p3titles = c("Iteration", "Week", "Data"),
  i.p4titles = c("Merged epidemics", "separation", "Iteration", "Week", "Data"),
  i.p5titles = c("Series separated in artificial seasons",
    "Series separated and MEM epidemics", "Week", "Data"),
  ...
)
}
\description{
divide a series in parts
}
\keyword{internal}
