% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformseries.multiple.R
\name{transformseries.multiple}
\alias{transformseries.multiple}
\title{divide a series in parts}
\usage{
transformseries.multiple(i.data, i.max.duration = NA, i.waves = NA,
  i.param.1 = 0.028, i.param.2 = 0.028, i.min.separation = 2,
  i.output = NA, i.force.loess = F)
}
\description{
divide a series in parts
}
\keyword{internal}
