% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformseries.multiple.R
\name{transformseries.multiple}
\alias{transformseries.multiple}
\title{divide a series in parts}
\usage{
transformseries.multiple(
  i.data,
  i.max.epidemic.duration = NA,
  i.max.season.duration = NA,
  i.waves = NA,
  i.param.1 = 3,
  i.param.2 = 2,
  i.min.separation = 1,
  i.output = NA,
  i.prefix = "Multiple waves",
  i.force.smooth = F,
  ...
)
}
\description{
divide a series in parts
}
\keyword{internal}
