% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformseries.twowaves.R
\name{transformseries.twowaves}
\alias{transformseries.twowaves}
\title{fills in missing values inside the season with smoothing regression}
\usage{
transformseries.twowaves(i.data, i.scale = 10000)
}
\description{
fills in missing values inside the season with smoothing regression
}
\keyword{internal}
