% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformseries.twowaves.R
\name{transformseries.twowaves}
\alias{transformseries.twowaves}
\title{adjust a mixed model of two normal distributions}
\usage{
transformseries.twowaves(i.data, i.scale = 10000)
}
\description{
adjust a mixed model of two normal distributions
}
\keyword{internal}
